backend/requirements.txt
azure-functions
pandas
scikit-learn
lightgbm
azure-storage-blob